refactor(eth): rpc/types -> rpc
