+++ eth/rpc/block_test.go
@@ -1,4 +1,4 @@
-package types
+package rpc
 
 import (
 	"fmt"
